add task, optimize performance

+++ routes/web.php
@@ -40,38 +40,48 @@ Route::group(['prefix' => LaravelLocalization::setLocale(),
         Route::get('/sellers', [App\Http\Controllers\DashboardController::class, 'sellers'])->name('sellers');
 
 
-        Route::get('/mybalance', [App\Http\Controllers\OrdersController::class, 'mybalance'])->name('mybalance');
-        Route::get('/states', [App\Http\Controllers\AreaController::class, 'states'])->name('states');
-        Route::post('/states/toggle/{id}', [\App\Http\Controllers\AreaController::class,'toggle']);
+
+        Route::get('/states', [App\Http\Controllers\Admin\AreaController::class, 'states'])->name('states');
+        Route::post('/states/toggle/{id}', [\App\Http\Controllers\Admin\AreaController::class,'toggle']);
         Route::get('/export/orders/en', [App\Http\Controllers\ExcelController::class,'exportOrdersEN'])->name('export.orders.en');
         Route::get('/export/orders/ar', [App\Http\Controllers\ExcelController::class,'exportOrdersAR'])->name('export.orders.ar');
 
-        Route::get('roles/permissions',[\App\Http\Controllers\RolesController::class,'permissions'])->name('roles.permissions');
-        Route::post('roles/permissions',[\App\Http\Controllers\RolesController::class,'permissionsCreate']);
-        Route::delete('permission/delete/{id}',[\App\Http\Controllers\RolesController::class,'permissionsDelete'])->name('permission.delete');
+        Route::get('roles/permissions',[\App\Http\Controllers\Admin\RolesController::class,'permissions'])->name('roles.permissions');
+        Route::post('roles/permissions',[\App\Http\Controllers\Admin\RolesController::class,'permissionsCreate']);
+        Route::delete('permission/delete/{id}',[\App\Http\Controllers\Admin\RolesController::class,'permissionsDelete'])->name('permission.delete');
 
-        Route::get('branches/assign/{id}',[\App\Http\Controllers\BranchController::class,'assign'])->name('branches.assign');
-        Route::post('branches/assign/{id}',[\App\Http\Controllers\BranchController::class,'assignGo'])->name('branches.assign');
+        Route::get('branches/assign/{id}',[\App\Http\Controllers\Admin\BranchController::class,'assign'])->name('branches.assign');
+        Route::post('branches/assign/{id}',[\App\Http\Controllers\Admin\BranchController::class,'assignGo'])->name('branches.assign');
 
-        Route::get('orders/assign',[\App\Http\Controllers\OrdersController::class,'assign'])->name('orders.assign');
-        Route::post('orders/assign',[\App\Http\Controllers\OrdersController::class,'assignGo'])->name('orders.assign');
+        Route::get('orders/assign',[\App\Http\Controllers\DashboardController::class,'assign'])->name('orders.assign');
+        Route::post('orders/assign',[\App\Http\Controllers\DashboardController::class,'assignGo'])->name('orders.assign');
+
+    Route::get('inventory',[\App\Http\Controllers\SellerController::class,'inventory'])->name('orders.inventory');
+    Route::get('orders/inline',[\App\Http\Controllers\SellerController::class,'inline'])->name('orders.inline');
+    Route::post('orders/inline/{id}',[\App\Http\Controllers\SellerController::class,'inlinego'])->name('orders.inlinego');
+    Route::post('orders/wait/{id}',[\App\Http\Controllers\SellerController::class,'wait'])->name('orders.wait');
+    Route::post('orders/pickup',[\App\Http\Controllers\SellerController::class,'pickup'])->name('orders.pickup');
+    Route::get('/mybalance', [App\Http\Controllers\SellerController::class, 'mybalance'])->name('mybalance');
+    Route::get('myorders',[\App\Http\Controllers\SellerController::class,'myorders']);
 
     Route::get('my-orders',[\App\Http\Controllers\DeliveryController::class,'myorders'])->name('myorders');
+
     Route::get('orders/status/{id}',[\App\Http\Controllers\DeliveryController::class,'status'])->name('orders.status');
     Route::post('orders/status/{id}/change',[\App\Http\Controllers\DeliveryController::class,'changeStatus'])->name('orders.changeStatus');
 
-        Route::resource('users',\App\Http\Controllers\UserController::class);
-        Route::resource('roles',\App\Http\Controllers\RolesController::class);
-        Route::resource('branches',\App\Http\Controllers\BranchController::class);
-        Route::resource('areas',\App\Http\Controllers\AreaController::class);
-        Route::resource('zones',\App\Http\Controllers\ZoneController::class);
+        Route::resource('users',\App\Http\Controllers\Admin\UserController::class);
+        Route::resource('roles',\App\Http\Controllers\Admin\RolesController::class);
+        Route::resource('branches',\App\Http\Controllers\Admin\BranchController::class);
+        Route::resource('areas',\App\Http\Controllers\Admin\AreaController::class);
+        Route::resource('zones',\App\Http\Controllers\Admin\ZoneController::class);
         Route::resource('orders',\App\Http\Controllers\OrdersController::class);
         Route::resource('receipts', App\Http\Controllers\ReceiptController::class);
+        Route::resource('tasks', App\Http\Controllers\TaskController::class);
 
-        Route::get('/e', [App\Http\Controllers\MailController::class, 'index'])->name('email');
-        Route::post('/e/s', [App\Http\Controllers\MailController::class, 'send'])->name('email.send');
-        Route::post('/e', [App\Http\Controllers\MailController::class, 'index'])->name('emailto');
-        Route::get('/welcome-email', [App\Http\Controllers\MailController::class, 'welcomeMail'])->name('welcome.email');
+        Route::get('/e', [App\Http\Controllers\Admin\MailController::class, 'index'])->name('email');
+        Route::post('/e/s', [App\Http\Controllers\Admin\MailController::class, 'send'])->name('email.send');
+        Route::post('/e', [App\Http\Controllers\Admin\MailController::class, 'index'])->name('emailto');
+        Route::get('/welcome-email', [App\Http\Controllers\Admin\MailController::class, 'welcomeMail'])->name('welcome.email');
 
     });
